Dual length, asterisks on patterns, regex change, sort by length then INI position
This branch is the closest to `get_browser` that I’ve been able to do
so far. It results in only 21 differences:

https://www.diffchecker.com/Qt36Y0Mq

As for performance, it’s also slower than other attempts I’ve done so
far, and I assume this is mostly because we have to look at every
suitable pattern (have to break out the tab separated groupings) to get
the position from the browscap.ini file which I’m storing with the
pattern.

 * Original: 26.269s
 * This branch: 62.481s
 * `get_browser`: 110.666s

Still faster than `get_browser` though, even on PHP 7.0.15. ;)