Some optimizations for performance
This gets the time taken to process the browscap test suite down to 51
seconds (the biggest win in here I believe is the bump to a 3 character
subkey rather than 2, which goes well with the increase in possible
variants from previous commit).

This is still slower than the 27 seconds it takes the unmodified
`browscap-php` but does bring it more in line with how `get_browser`
parses (which is still much slower than even this branch).

The diff is still the same as before (just 16 agents are different, all
of which are the result of the digit replacement, one way or another).